7zpp: fixed signedness warning
vector<T>::size() returns size_t, which is unsigned, whereas int is
signed.

+++ endless/src/7zip-cpp/7zpp/UsefulFunctions.cpp
@@ -215,7 +215,7 @@ namespace SevenZip
 		myAvailableFormats.push_back(CompressionFormat::SquashFS);
 
 		// Check each format for one that works
-		for (int i = 0; i < myAvailableFormats.size(); i++)
+		for (size_t i = 0; i < myAvailableFormats.size(); i++)
 		{
 			archiveCompressionFormat = myAvailableFormats[i];
